Refactor (part 2)
