+++ rest-service/manager_rest/upload_manager/manager.py
@@ -1,16 +1,11 @@
 import os
-import json
 import tarfile
 import uuid
 
-import wagon
-import yaml
 import shutil
 import zipfile
 import tempfile
-import requests
 
-from setuptools import archive_util
 from flask import request, current_app
 
 from cloudify.models_states import BlueprintUploadState
@@ -23,7 +18,7 @@ from manager_rest.constants import (FILE_SERVER_PLUGINS_FOLDER,
 from manager_rest.dsl_back_compat import create_bc_plugin_yaml
 from manager_rest.archiving import get_archive_type
 from manager_rest.storage.models import Blueprint
-from manager_rest import config, chunked, manager_exceptions
+from manager_rest import config, manager_exceptions
 from manager_rest.utils import (mkdirs,
                                 current_tenant,
                                 unzip,
@@ -31,178 +26,24 @@ from manager_rest.utils import (mkdirs,
                                 remove)
 from manager_rest.resource_manager import get_resource_manager
 from manager_rest.constants import (SUPPORTED_ARCHIVE_TYPES)
+from manager_rest.upload_manager.utils import (
+    base_archive_filename,
+    extract_file_to_file_server,
+    is_wagon_file,
+    load_plugin_extras,
+    load_plugin_package_json,
+    save_file_content,
+    save_file_locally_and_extract_inputs,
+    unpack_caravan,
+    zip_dir,
+)
 
 
 UPLOADING_FOLDER_NAME = '.uploading'
 
 
-def _zip_dir(dir_to_zip, target_zip_path):
-    zipf = zipfile.ZipFile(target_zip_path, 'w', zipfile.ZIP_DEFLATED)
-    try:
-        plugin_dir_base_name = os.path.basename(dir_to_zip)
-        rootlen = len(dir_to_zip) - len(plugin_dir_base_name)
-        for base, dirs, files in os.walk(dir_to_zip):
-            for entry in files:
-                fn = os.path.join(base, entry)
-                zipf.write(fn, fn[rootlen:])
-    finally:
-        zipf.close()
-
-
-def _extract_file_to_file_server(archive_path, destination_root):
-    """
-    Extracting a package.
-
-    :param destination_root: the root destination for the unzipped archive
-    :param archive_path: the archive path
-    :return: the full path for the extracted archive
-    """
-    # extract application to file server
-    tempdir = tempfile.mkdtemp('-blueprint-submit')
-    try:
-        try:
-            archive_util.unpack_archive(archive_path, tempdir)
-        except archive_util.UnrecognizedFormat:
-            raise manager_exceptions.BadParametersError(
-                'Blueprint archive is of an unrecognized format. '
-                'Supported formats are: {0}'
-                .format(SUPPORTED_ARCHIVE_TYPES))
-        archive_file_list = os.listdir(tempdir)
-        if len(archive_file_list) != 1 or not os.path.isdir(
-                os.path.join(tempdir, archive_file_list[0])):
-            raise manager_exceptions.BadParametersError(
-                'archive must contain exactly 1 directory')
-        application_dir_base_name = archive_file_list[0]
-        # generating temporary unique name for app dir, to allow multiple
-        # uploads of apps with the same name (as it appears in the file
-        # system, not the app name field inside the blueprint.
-        # the latter is guaranteed to be unique).
-        generated_app_dir_name = '{0}-{1}'.format(
-            application_dir_base_name, uuid.uuid4())
-        temp_application_dir = os.path.join(tempdir,
-                                            application_dir_base_name)
-        temp_application_target_dir = os.path.join(tempdir,
-                                                   generated_app_dir_name)
-        shutil.move(temp_application_dir, temp_application_target_dir)
-        shutil.move(temp_application_target_dir, destination_root)
-        return generated_app_dir_name
-    finally:
-        shutil.rmtree(tempdir)
-
-
-def _save_file_from_url(archive_target_path, url, data_type):
-    if request.data or \
-            'Transfer-Encoding' in request.headers or \
-            'blueprint_archive' in request.files:
-        raise manager_exceptions.BadParametersError(
-            "Can pass {0} as only one of: URL via query parameters, "
-            "request body, multi-form or chunked.".format(data_type))
-    try:
-        with requests.get(url, stream=True, timeout=(5, None)) as resp:
-            resp.raise_for_status()
-            with open(archive_target_path, 'wb') as f:
-                for chunk in resp.iter_content(chunk_size=8192):
-                    if chunk:
-                        f.write(chunk)
-    except requests.exceptions.RequestException as e:
-        raise manager_exceptions.BadParametersError(
-            "Cannot fetch {0}: {1}".format(url, e))
-
-
-def _save_file_content(archive_target_path, data_type):
-    if 'blueprint_archive' in request.files:
-        raise manager_exceptions.BadParametersError(
-            "Can't pass {0} both as URL via request body and multi-form"
-            .format(data_type))
-    uploaded_file_data = request.data
-    with open(archive_target_path, 'wb') as f:
-        f.write(uploaded_file_data)
-
-
-def _save_files_multipart(archive_target_path):
-    inputs = {}
-    for file_key in request.files:
-        if file_key == 'inputs':
-            content = request.files[file_key]
-            # The file is a binary
-            if 'application' in content.content_type:
-                content_payload = _save_bytes(content)
-                # Handling yaml
-                if content.content_type == 'application/octet-stream':
-                    inputs = yaml.safe_load(content_payload)
-                # Handling json
-                elif content.content_type == 'application/json':
-                    inputs = json.load(content_payload)
-            # The file is raw json
-            elif 'text' in content.content_type:
-                inputs = json.load(content)
-        elif file_key == 'blueprint_archive':
-            _save_bytes(request.files[file_key],
-                        archive_target_path)
-    return inputs
-
-
-def _save_bytes(content, target_path=None):
-    """
-    content should support read() function if target isn't supplied,
-    string rep is returned
-
-    :param content:
-    :param target_path:
-    :return:
-    """
-    if not target_path:
-        return content.getvalue().decode("utf-8")
-    else:
-        with open(target_path, 'wb') as f:
-            f.write(content.read())
-
-
-def _save_file_locally_and_extract_inputs(archive_target_path,
-                                          url_key,
-                                          data_type='unknown'):
-    """
-    Retrieves the file specified by the request to the local machine.
-
-    :param archive_target_path: the target of the archive
-    :param data_type: the kind of the data (e.g. 'blueprint')
-    :param url_key: if the data is passed as a url to an online resource,
-    the url_key specifies what header points to the requested url.
-    :return: None
-    """
-    inputs = {}
-    # Handling importing blueprint through url
-    if url_key in request.args:
-        _save_file_from_url(archive_target_path,
-                            request.args[url_key],
-                            data_type)
-    # handle receiving chunked blueprint
-    elif 'Transfer-Encoding' in request.headers:
-        _save_file_from_chunks(archive_target_path, data_type)
-    # handler receiving entire content through data
-    elif request.data:
-        _save_file_content(archive_target_path, data_type)
-
-    # handle inputs from form-data (for both the blueprint and inputs
-    # in body in form-data format)
-    if request.files:
-        inputs = _save_files_multipart(archive_target_path)
-
-    return inputs
-
-
-def _save_file_from_chunks(archive_target_path, data_type):
-    if request.data or 'blueprint_archive' in request.files:
-        raise manager_exceptions.BadParametersError(
-            "Can pass {0} as only one of: request body, multi-form or "
-            "chunked.".format(data_type))
-    with open(archive_target_path, 'w') as f:
-        for buffered_chunked in chunked.decode(request.input_stream):
-            f.write(buffered_chunked)
-
-
 def _do_upload_snapshot(snapshot_id, upload_path):
-    _save_file_locally_and_extract_inputs(
+    save_file_locally_and_extract_inputs(
         upload_path,
         'snapshot_archive_url',
         'snapshot',
@@ -233,7 +74,7 @@ def upload_snapshot(snapshot_id):
 
 
 def _do_upload_blueprint(blueprint_id, upload_path):
-    _save_file_locally_and_extract_inputs(
+    save_file_locally_and_extract_inputs(
         upload_path,
         None,
         'blueprint')
@@ -281,7 +122,7 @@ def cleanup_blueprint_archive_from_file_server(blueprint_id, tenant):
 
 def update_blueprint_icon_file(self, tenant_name, blueprint_id):
     icon_tmp_path = tempfile.mktemp()
-    _save_file_content(icon_tmp_path, 'blueprint_icon')
+    save_file_content(icon_tmp_path, 'blueprint_icon')
     _set_blueprints_icon(tenant_name, blueprint_id, icon_tmp_path)
     remove(icon_tmp_path)
     _update_blueprint_archive(tenant_name, blueprint_id)
@@ -319,7 +160,7 @@ def _update_blueprint_archive(tenant_name, blueprint_id):
     # Filename will be like [BLUEPRINT_ID].tar.gz or [BLUEPRINT_ID].zip
     archive_filename = [fn for fn in os.listdir(archive_dir)
                         if fn.startswith(blueprint_id)][0]
-    base_filename = _base_archive_filename(archive_filename)
+    base_filename = base_archive_filename(archive_filename)
     orig_archive_path = os.path.join(archive_dir, archive_filename)
     new_archive_path = os.path.join(archive_dir, f'{base_filename}.tar.gz')
     with tempfile.TemporaryDirectory(dir=file_server_root) as tmpdir:
@@ -368,8 +209,8 @@ def extract_blueprint_archive_to_file_server(blueprint_id, tenant):
         sm.update(blueprint)
         raise manager_exceptions.NotFoundError(error_msg)
     try:
-        app_dir = _extract_file_to_file_server(local_file_path,
-                                               file_server_root)
+        app_dir = extract_file_to_file_server(local_file_path,
+                                              file_server_root)
     except Exception as e:
         blueprint = sm.get(Blueprint, blueprint_id)
         blueprint.state = \
@@ -413,69 +254,7 @@ def _process_blueprint_plugins(file_server_root, blueprint_id):
     for plugin_dir in plugins:
         final_zip_name = '{0}.zip'.format(os.path.basename(plugin_dir))
         target_zip_path = os.path.join(plugins_directory, final_zip_name)
-        _zip_dir(plugin_dir, target_zip_path)
-
-
-def _load_plugin_package_json(wagon_source):
-    try:
-        return wagon.show(wagon_source)
-    except (wagon.WagonError, tarfile.ReadError, zipfile.BadZipFile) as e:
-        raise manager_exceptions.InvalidPluginError(
-            'The provided wagon archive can not be read.\n{0}: {1}'
-            .format(type(e).__name__, e))
-
-
-def _is_wagon_file(file_path):
-    try:
-        _load_plugin_package_json(file_path)
-    except Exception:
-        return False
-    else:
-        return True
-
-
-def _load_plugin_extras(filenames):
-    filename = _choose_plugin_yaml(filenames)
-    if not filename:
-        return {}
-
-    with open(filename, 'r') as fh:
-        try:
-            plugin_yaml = yaml.safe_load(fh) or {}
-        except yaml.YAMLError as e:
-            raise manager_exceptions.InvalidPluginError(
-                f"The provided plugin's description ({filename}) "
-                f"can not be read.\n{e}")
-    return {
-        'blueprint_labels': _retrieve_labels(
-            plugin_yaml.get('blueprint_labels')),
-        'labels': _retrieve_labels(plugin_yaml.get('labels')),
-        'resource_tags': plugin_yaml.get('resource_tags'),
-    }
-
-
-def _choose_plugin_yaml(filenames):
-    if not filenames:
-        return None
-    for fn in filenames:
-        if 'plugin.' in fn.lower():
-            return fn
-    return filenames[0]
-
-
-def _unpack_caravan(path, directory):
-    if not tarfile.is_tarfile(path):
-        return None
-
-    try:
-        with tarfile.open(path) as caravan:
-            root_path = caravan.getmembers()[0].path
-            caravan.extractall(directory)
-    except tarfile.ReadError:
-        return None
-
-    with open(os.path.join(directory, root_path, 'METADATA')) as f:
-        return json.load(f), root_path
+        zip_dir(plugin_dir, target_zip_path)
 
 
 def _store_plugin(plugin_id, wagon_path, yaml_paths):
@@ -486,8 +265,8 @@ def _store_plugin(plugin_id, wagon_path, yaml_paths):
         'resource_tags': None,
     }
 
-    wagon_info.update(_load_plugin_package_json(wagon_path))
-    wagon_info.update(_load_plugin_extras(yaml_paths))
+    wagon_info.update(load_plugin_package_json(wagon_path))
+    wagon_info.update(load_plugin_extras(yaml_paths))
 
     plugin_dir = os.path.dirname(wagon_path)
     if yaml_paths:
@@ -531,7 +310,7 @@ def is_caravan(path):
 
 def _do_upload_plugin(data_id, plugin_dir):
     archive_target_path = os.path.join(plugin_dir, 'plugin')
-    _save_file_locally_and_extract_inputs(
+    save_file_locally_and_extract_inputs(
         archive_target_path,
         'plugin_archive_url',
         'plugin',
@@ -539,7 +318,7 @@ def _do_upload_plugin(data_id, plugin_dir):
 
     plugins = []
     if is_caravan(archive_target_path):
-        metadata, root_path = _unpack_caravan(archive_target_path, plugin_dir)
+        metadata, root_path = unpack_caravan(archive_target_path, plugin_dir)
         for wagon_path, yaml_path in metadata.items():
             wagon_path = os.path.join(
                 plugin_dir,
@@ -557,7 +336,7 @@ def _do_upload_plugin(data_id, plugin_dir):
                 [yaml_path],
             )
             plugins.append(plugin_desc)
-    elif _is_wagon_file(archive_target_path):
+    elif is_wagon_file(archive_target_path):
         plugins = [(data_id, archive_target_path, [])]
     elif zipfile.is_zipfile(archive_target_path):
         plugin_dir = os.path.dirname(archive_target_path)
@@ -599,36 +378,3 @@ def upload_plugin(data_id=None, **kwargs):
         return _do_upload_plugin(data_id, upload_path)
     finally:
         shutil.rmtree(upload_path, ignore_errors=True)
-
-
-def _retrieve_labels(labels):
-    if not labels:
-        return labels
-    try:
-        flattened = _flatten_labels_dict(labels)
-    except AttributeError:
-        raise manager_exceptions.InvalidPluginError(
-            f"Invalid labels: {labels}")
-    for k, v in flattened.items():
-        if not isinstance(v, list):
-            raise manager_exceptions.InvalidPluginError(
-                f"Invalid labels: {labels}")
-    return flattened
-
-
-def _flatten_labels_dict(labels):
-    """Flatten labels dictionary by dropping 'values' key."""
-    flattened = {}
-    for k, v in labels.items():
-        flattened[k] = v.get('values')
-    return flattened
-
-
-def _base_archive_filename(archive_filename):
-    while True:
-        filename, _, ext = archive_filename.rpartition('.')
-        if ext in ['tar', 'tgz', 'zip']:
-            return filename
-        if ext in ['bz2', 'gz', 'lzma', 'xz']:
-            continue
-        return archive_filename
